d/rules: Don't display the sample list of "Editor's Choice" apps
Upstream no longer recommends using this list and does not maintain it,
and we don't want to keep recommending apps that are dead upstream
such as gnome-dictionary and gnome-photos. Ideally we should provide
our own curated list for forky, but until we can do that, a minimal
fix is to not display it.

Helps: #1100084